requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
APScheduler==3.9.1
asgiref==3.5.2
astroid==2.12.13
async-timeout==4.0.2
attrs==22.1.0
Authlib==1.1.0
black==22.8.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cryptography==38.0.1
dill==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-APScheduler==1.12.4
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.1
greenlet==1.1.3.post0
gunicorn==20.1.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==4.12.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keybert==0.7.0
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.2
mypy-extensions==0.4.3
nltk==3.7
numpy==1.23.4
packaging==21.3
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.5.2
pycparser==2.21
Pygments==2.13.0
pylint==2.15.6
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rich==12.6.0
scikit-learn==1.1.3
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
tomlkit==0.11.6
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
